Fix merge and tween only if not tweening

+++ Engine/Component/ComponentAnimation.h
@@ -45,7 +45,7 @@ public:
 	ENGINE_API void ActiveAnimation(const std::string & trigger);
 	ENGINE_API bool IsOnState(const std::string & trigger);
 	ENGINE_API float GetCurrentClipPercentatge() const;
-	ENGINE_API int GetTotalAnimationTime() const;
+	ENGINE_API float GetTotalAnimationTime() const;
 	ENGINE_API void SetAnimationSpeed(float speed) const;
 
 	void Update() override;
